Extract libinput::create_handler method
