+++ windows-gaming-driver/src/driver/libinput.rs
@@ -1,16 +1,24 @@
 use std::io;
+use std::iter;
+use std::rc::Rc;
 use std::cell::RefCell;
 
 use tokio_core::reactor::{Handle, PollEvented};
-use futures::{Async, Poll, Future};
+use futures::{Async, Poll, Future, Stream, Sink};
 use futures::unsync::mpsc::{UnboundedSender, UnboundedReceiver, self};
 use input::{Libinput, LibinputInterface, Device};
-use input::event::Event;
+use input::event::{Event, KeyboardEvent, PointerEvent};
+use input::event::pointer::{Axis, ButtonState};
+use input::event::keyboard::{KeyState, KeyboardEventTrait};
 use libc::{self, c_char, c_int, c_ulong, c_void};
 use libudev::{Result as UdevResult, Context, Enumerator};
 
 use super::my_io::MyIo;
 use config::{UsbDevice, UsbBinding, UsbPort, UsbId};
+use driver::controller::Controller;
+use driver::hotkeys::{KeyboardState, KeyResolution};
+use driver::monitor::{QmpCommand, InputEvent, InputButton, KeyValue};
+use driver::hotkeys::KeyBinding;
 
 const EVIOCGRAB: c_ulong = 1074021776;
 
@@ -24,6 +32,12 @@ unsafe extern "C" fn do_close(fd: c_int, _: *mut c_void) {
     libc::ioctl(fd, EVIOCGRAB, 0);
     libc::close(fd);
 }
+
+const BTN_LEFT: u32 = 0x110;
+const BTN_RIGHT: u32 = 0x111;
+const BTN_MIDDLE: u32 = 0x112;
+const BTN_SIDE: u32 = 0x113;
+const BTN_EXTRA: u32 = 0x114;
 
 pub struct Input {
     usb_devs: Vec<UsbDevice>,
@@ -119,12 +133,90 @@ impl<'a> Future for InputListener<'a> {
 
         // lmao what a quality api
         while let Some(e) = p.li.next() {
-            p.sender.send(e).unwrap();
+            (&p.sender).send(e).unwrap();
         }
 
         p.io.need_read();
         Ok(Async::NotReady)
     }
+}
+
+pub fn create_handler<'a>(input_events: UnboundedReceiver<Event>, hotkey_bindings: &'a [KeyBinding],
+                      controller: Rc<RefCell<Controller>>, monitor_sender: UnboundedSender<QmpCommand>)
+            -> Box<Future<Item = (), Error = io::Error> + 'a> {
+    let mut keyboard_state = KeyboardState::new(hotkey_bindings);
+    let input_handler = input_events.filter_map(move |event| {
+        Some(match event {
+            Event::Pointer(PointerEvent::Motion(m)) =>
+                QmpCommand::InputSendEvent { events: vec![
+                    InputEvent::Rel { axis: "x", value: m.dx() as u32 },
+                    InputEvent::Rel { axis: "y", value: m.dy() as u32 },
+                ]},
+            Event::Pointer(PointerEvent::Button(b)) =>
+                QmpCommand::InputSendEvent {
+                    events: vec![InputEvent::Btn {
+                        down: b.button_state() == ButtonState::Pressed,
+                        button: match b.button() {
+                            BTN_LEFT => InputButton::Left,
+                            BTN_RIGHT => InputButton::Right,
+                            BTN_MIDDLE => InputButton::Middle,
+                            BTN_SIDE => InputButton::Side,
+                            BTN_EXTRA => InputButton::Extra,
+                            b => {
+                                warn!("Unknown mouse button {}", b);
+                                return None;
+                            }
+                        }
+                    }]
+                },
+            Event::Pointer(PointerEvent::Axis(ref a)) if a.has_axis(Axis::Vertical) => {
+                let steps = a.axis_value_discrete(Axis::Vertical).map(|x| x as i32).unwrap_or(0);
+                if steps == 0 {
+                    // stop event, ignore
+                    return None;
+                }
+
+                let direction = if steps > 0 {
+                    InputButton::WheelDown
+                } else {
+                    InputButton::WheelUp
+                };
+
+                QmpCommand::InputSendEvent {
+                    events: iter::repeat(direction).take(steps.abs() as usize).flat_map(|b| vec![
+                        InputEvent::Btn { down: true, button: b },
+                        InputEvent::Btn { down: false, button: b },
+                    ]).collect(),
+                }
+            },
+            Event::Keyboard(KeyboardEvent::Key(k)) => {
+                let down = k.key_state() == KeyState::Pressed;
+                let KeyResolution { hotkeys, qcode } = match keyboard_state.input_linux(k.key(), down) {
+                    Some(x) => x,
+                    None => return None,
+                };
+
+                for hk in hotkeys {
+                    controller.borrow_mut().ga_hotkey(hk as u32);
+                }
+
+                match qcode {
+                    Some(qcode) => QmpCommand::InputSendEvent {
+                        events: vec![ InputEvent::Key {
+                            down,
+                            key: KeyValue::Qcode(qcode),
+                        }]
+                    },
+                    None => return None
+                }
+            }
+            event => {
+                info!("Unhandled input event {:?}", event);
+                return None;
+            }
+        })
+    }).forward(monitor_sender.sink_map_err(|_| ())).then(|_| Ok(()));
+    Box::new(input_handler)
 }
 
 /*
