+++ windows-gaming-driver/src/driver/mod.rs
@@ -21,20 +21,14 @@ use std::os::unix::fs::PermissionsExt;
 use std::path::Path;
 
 use tokio_core::reactor::Core;
-use futures::{Future, Stream, Sink, future};
+use futures::{Future, Stream, future};
 
 use driver::controller::Controller;
 use config::Config;
 use driver::signalfd::{SignalFd, signal};
-use driver::hotkeys::{KeyboardState, KeyResolution};
 use self::monitor::Monitor;
 use self::clientpipe::Clientpipe;
-
-const BTN_LEFT: u32 = 0x110;
-const BTN_RIGHT: u32 = 0x111;
-const BTN_MIDDLE: u32 = 0x112;
-const BTN_SIDE: u32 = 0x113;
-const BTN_EXTRA: u32 = 0x114;
+use self::libinput::Input;
 
 pub fn run(cfg: &Config, tmp: &Path, data: &Path) {
     let _ = fs::remove_dir_all(tmp); // may fail - we dont care
@@ -80,7 +74,7 @@ pub fn run(cfg: &Config, tmp: &Path, data: &Path) {
     let mut monitor = Monitor::new(monitor_stream, &handle);
     let mut clientpipe = Clientpipe::new(clientpipe_stream, &handle);
 
-    let (mut input, input_events) = libinput::Input::new(&handle, cfg.machine.usb_devices.clone());
+    let (mut input, input_events) = Input::new(&handle, cfg.machine.usb_devices.clone(),);
     input.suspend();
     let input = Rc::new(RefCell::new(input));
 
@@ -95,83 +89,8 @@ pub fn run(cfg: &Config, tmp: &Path, data: &Path) {
     let ref input_ref = *input;
     let input_listener = libinput::InputListener(input_ref);
     let hotkey_bindings: Vec<_> = cfg.machine.hotkeys.iter().map(|x| x.key.clone()).collect();
-    let mut keyboard_state = KeyboardState::new(&hotkey_bindings);
-    let input_handler = input_events.filter_map(|event| {
-        use self::monitor::*;
-        use std::iter;
-        use input::event::{Event, KeyboardEvent, PointerEvent};
-        use input::event::pointer::{Axis, ButtonState};
-        use input::event::keyboard::{KeyState, KeyboardEventTrait};
-        Some(match event {
-            Event::Pointer(PointerEvent::Motion(m)) =>
-                QmpCommand::InputSendEvent { events: vec![
-                    InputEvent::Rel { axis: "x", value: m.dx() as u32 },
-                    InputEvent::Rel { axis: "y", value: m.dy() as u32 },
-                ]},
-            Event::Pointer(PointerEvent::Button(b)) =>
-                QmpCommand::InputSendEvent {
-                    events: vec![InputEvent::Btn {
-                        down: b.button_state() == ButtonState::Pressed,
-                        button: match b.button() {
-                            BTN_LEFT => InputButton::Left,
-                            BTN_RIGHT => InputButton::Right,
-                            BTN_MIDDLE => InputButton::Middle,
-                            BTN_SIDE => InputButton::Side,
-                            BTN_EXTRA => InputButton::Extra,
-                            b => {
-                                warn!("Unknown mouse button {}", b);
-                                return None;
-                            }
-                        }
-                    }]
-                },
-            Event::Pointer(PointerEvent::Axis(ref a)) if a.has_axis(Axis::Vertical) => {
-                let steps = a.axis_value_discrete(Axis::Vertical).map(|x| x as i32).unwrap_or(0);
-                if steps == 0 {
-                    // stop event, ignore
-                    return None;
-                }
-
-                let direction = if steps > 0 {
-                    InputButton::WheelDown
-                } else {
-                    InputButton::WheelUp
-                };
-
-                QmpCommand::InputSendEvent {
-                    events: iter::repeat(direction).take(steps.abs() as usize).flat_map(|b| vec![
-                        InputEvent::Btn { down: true, button: b },
-                        InputEvent::Btn { down: false, button: b },
-                    ]).collect(),
-                }
-            },
-            Event::Keyboard(KeyboardEvent::Key(k)) => {
-                let down = k.key_state() == KeyState::Pressed;
-                let KeyResolution { hotkeys, qcode } = match keyboard_state.input_linux(k.key(), down) {
-                    Some(x) => x,
-                    None => return None,
-                };
-
-                for hk in hotkeys {
-                    controller.borrow_mut().ga_hotkey(hk as u32);
-                }
-
-                match qcode {
-                    Some(qcode) => QmpCommand::InputSendEvent {
-                        events: vec![ InputEvent::Key {
-                            down,
-                            key: KeyValue::Qcode(qcode),
-                        }]
-                    },
-                    None => return None
-                }
-            }
-            event => {
-                info!("Unhandled input event {:?}", event);
-                return None;
-            }
-        })
-    }).forward(monitor_sender.sink_map_err(|_| ())).then(|_| Ok(()));
+    let input_handler = libinput::create_handler(input_events, &hotkey_bindings, controller.clone(),
+                                                 monitor_sender);
 
     let control_handler = control::create(control_socket, &handle, controller.clone());
 
@@ -190,7 +109,7 @@ pub fn run(cfg: &Config, tmp: &Path, data: &Path) {
         monitor.take_sender(),
         Box::new(catch_sigterm),
         Box::new(input_listener),
-        Box::new(input_handler),
+        input_handler,
     ]).map(|_| ());
 
     match core.run(qemu.select(joined)) {
